requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.0.2
SQLAlchemy==2.0.27
beautifulsoup4==4.12.3
requests==2.31.0
schedule==1.2.1 